api: prompt user to reset password during OAuth2 authorization
This password reset mechanism differs a bit from the previous setup
where the webui handled password resets. In the case of webui, the user
was first logged in and then he was forced to reset his password. This
has led many users to lock-out their accounts when they failed to reset
the password in time or when they accidentally clicked the login button
twice.

Now the user is prompted to reset his password before he is actually
logged in. If the user fails to reset his password, he can still
authenticate and reset it later. Having the password reset facility
included within the OAuth2 authorization process also means that the
password will be reset to whichever service the users wishes to log in,
before it had to be the webui, otherwise he would get locked out.

User lock-out facility is now largely unused, it will have an effect
only when the user will use the token authentication mechanism on his
first login.

+++ api/lib/vpsadmin/api/operations/authentication/totp.rb
@@ -2,7 +2,7 @@ require 'vpsadmin/api/operations/base'
 
 module VpsAdmin::API
   class Operations::Authentication::Totp < Operations::Base
-    Result = Struct.new(:user, :auth_token, :recovery_device) do
+    Result = Struct.new(:user, :auth_token, :recovery_device, :reset_password) do
       def authenticated?
         !auth_token.nil?
       end
@@ -10,6 +10,8 @@ module VpsAdmin::API
       def used_recovery_code?
         !recovery_device.nil?
       end
+
+      alias_method :reset_password?, :reset_password
     end
 
     # @param token [String]
@@ -18,6 +20,7 @@ module VpsAdmin::API
     def run(token, code)
       auth_token = ::AuthToken.joins(:token).includes(:token, :user).find_by(
         tokens: {token: token},
+        purpose: 'totp',
       )
 
       if auth_token.nil? || !auth_token.token_valid?
@@ -41,8 +44,20 @@ module VpsAdmin::API
             dev.update!(enabled: false)
           end
 
-          auth_token.destroy!
-          return Result.new(user, auth_token, last_verification_at ? nil : dev)
+          reset_password = user.password_reset
+
+          if reset_password
+            auth_token.update!(purpose: 'reset_password')
+          else
+            auth_token.destroy!
+          end
+
+          return Result.new(
+            user,
+            auth_token,
+            last_verification_at ? nil : dev,
+            reset_password,
+          )
         end
       end
